refactor: reorganize src/ folder structure
- readers/ + core/input/ + core/services/parseInput + platform/web/ → input/
- controllers/ → hooks/
- presentation/ → ui/presentation/

Import paths updated throughout; all 172 tests pass.
