+++ src/input/buildInputData.js
@@ -11,15 +11,15 @@
  *   buildInputData(csvRows, trades)      → InputData
  */
 
-import { parseCSV }                                        from '../../readers/readCsv.js'
-import { parseTradesFromHtml }                             from '../domain/parser/parseTradesHtml.js'
-import { parseStatementInfo }                              from '../domain/parser/parseStatementInfo.js'
-import { parseInstruments }                                from '../domain/parser/parseInstruments.js'
-import { parseDividends, parseWithholdingTax }             from '../domain/parser/parseDividends.js'
-import { parseOpenPositions }                              from '../domain/parser/parseOpenPositions.js'
-import { parseCsvTrades }                                  from '../domain/parser/parseCsvTrades.js'
-import { parseInterest }                                   from '../domain/parser/parseInterest.js'
-import { parseTaxYear }                                    from '../domain/parser/parseTaxYear.js'
+import { parseCSV }                                        from './readCsv.js'
+import { parseTradesFromHtml }                             from '../core/domain/parser/parseTradesHtml.js'
+import { parseStatementInfo }                              from '../core/domain/parser/parseStatementInfo.js'
+import { parseInstruments }                                from '../core/domain/parser/parseInstruments.js'
+import { parseDividends, parseWithholdingTax }             from '../core/domain/parser/parseDividends.js'
+import { parseOpenPositions }                              from '../core/domain/parser/parseOpenPositions.js'
+import { parseCsvTrades }                                  from '../core/domain/parser/parseCsvTrades.js'
+import { parseInterest }                                   from '../core/domain/parser/parseInterest.js'
+import { parseTaxYear }                                    from '../core/domain/parser/parseTaxYear.js'
 import {
   validateCsvContent,
   validateHtmlContent,
