+++ src/ui/ResultTabs/HoldingsTab.jsx
@@ -3,7 +3,7 @@ import { Box, Typography } from '@mui/material'
 import { alpha } from '@mui/material/styles'
 import { ReceiptLongOutlined } from '@mui/icons-material'
 import DataTable from './DataTable.jsx'
-import { HoldingPresenter } from '../../presentation/HoldingPresenter.js'
+import { HoldingPresenter } from '../presentation/HoldingPresenter.js'
 import ThresholdWarning from '../ThresholdWarning.jsx'
 
 function addRowNumbers(rows) {
