+++ src/ui/ResultTabs/InterestTab.jsx
@@ -3,7 +3,7 @@ import { Box, Typography } from '@mui/material'
 import { alpha } from '@mui/material/styles'
 import { ReceiptLongOutlined } from '@mui/icons-material'
 import DataTable from './DataTable'
-import { InterestPresenter } from '../../presentation/InterestPresenter.js'
+import { InterestPresenter } from '../presentation/InterestPresenter.js'
 
 export default function InterestTab({ result }) {
   const { localCurrencyLabel, localCurrencyCode } = result
