+++ src/ui/ResultTabs/TaxSummary.jsx
@@ -2,7 +2,7 @@ import { t } from '../../localization/i18n.js'
 import { Box, Typography } from '@mui/material'
 import { InfoOutlined } from '@mui/icons-material'
 import { TaxFormSection, TaxRow } from './TaxFormSection.jsx'
-import { TradeSummaryPresenter } from '../../presentation/TradeSummaryPresenter.js'
+import { TradeSummaryPresenter } from '../presentation/TradeSummaryPresenter.js'
 
 
 export default function TaxSummary({ taxSummary, localCurrencyLabel = 'BGN' }) {
